+++ src/FAB.php
@@ -3,6 +3,10 @@
 namespace yii\fontawesome;
 
 /**
- * @deprecated Use FA instead
+ * FontAwesome Brands helper
+ * Use FAB::icon('icon-name')
  */
-class FAB extends FA {}
+class FAB extends FontAwesome
+{
+    public static string $cssPrefix = 'fa-brand';
+}