+++ src/FAD.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace yii\fontawesome;
-
-/**
- * @deprecated Use FA instead
- */
-class FAD extends FA {}