+++ src/FADL.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace yii\fontawesome;
+
+/**
+ * Duotone Light icons are not available in Free version!
+ */
+class FADL extends FA {}