+++ src/FADR.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace yii\fontawesome;
+
+/**
+ * Duotone Regular icons are not available in Free version!
+ */
+class FADR extends FA {}