+++ src/FADS.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace yii\fontawesome;
+
+/**
+ * Duotone Solid icons are not available in Free version!
+ */
+class FADS extends FA {}