+++ src/FAL.php
@@ -3,6 +3,6 @@
 namespace yii\fontawesome;
 
 /**
- * @deprecated Use FA instead
+ * Light icons are not available in Free version!
  */
 class FAL extends FA {}