+++ src/FASDL.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace yii\fontawesome;
+
+/**
+ * Shape Duotone Light icons are not available in Free version!
+ */
+class FASDL extends FA {}