+++ src/FASDR.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace yii\fontawesome;
+
+/**
+ * Shape Duotone Regular icons are not available in Free version!
+ */
+class FASDR extends FA {}