+++ src/FASDS.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace yii\fontawesome;
+
+/**
+ * Shape Duotone Solid icons are not available in Free version!
+ */
+class FASDS extends FA {}