+++ src/FASDT.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace yii\fontawesome;
+
+/**
+ * Shape Duotone Thin icons are not available in Free version!
+ */
+class FASDT extends FA {}