+++ src/FASL.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace yii\fontawesome;
+
+/**
+ * Shape Light icons are not available in Free version!
+ */
+class FASL extends FA {}