+++ src/FASR.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace yii\fontawesome;
+
+/**
+ * Shape Regular icons are not available in Free version!
+ */
+class FASR extends FA {}