+++ src/FASS.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace yii\fontawesome;
+
+/**
+ * Shape Solid icons are not available in Free version!
+ */
+class FASS extends FA {}