+++ src/FAST.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace yii\fontawesome;
+
+/**
+ * Shape Thin icons are not available in Free version!
+ */
+class FAST extends FA {}