+++ CHANGELOG.md
@@ -6,7 +6,11 @@ v6.7.1
 ------------------------
 - Bumped versions of PHP and Yii2 framework
 - Versions of package will be same as FontAwesome bundle source for more capability
-- Deprecated classes FAB, FAD, FAL, FAR, FAS, FAT are removed!!! They will be available only in PRO version of this package https://github.com/mylistryx/yii2-fontawesome-pro
+- Deprecated class FAB restored for Brands collection of icons
+- FAL, FAR, FAS, FAT not deprecated from this time, but they will always give Solid class for backward capability with PRO version! They will be available only in PRO version of this package https://github.com/mylistryx/yii2-fontawesome-pro
+- Added FADL, FADR, FADS, FADT helpers for capability with Duotone icons, but they still return Solid
+- Added FASL, FASR, FASS, FAST helpers for capability with Shape icons, but they still return Solid
+- Added FASDL, FASDR, FASDS, FASDT helpers for capability with Shape Duotone icons, but they still return Solid
 
 
 v2.0.0